Extract useData
