+++ client/sites/tools/logs/hooks/use-data.ts
@@ -1,0 +1,52 @@
+import {
+	LogType,
+	PHPLog,
+	ServerLog,
+	useSiteLogsQuery,
+} from 'calypso/data/hosting/use-site-logs-query';
+import { useSelector } from 'calypso/state';
+import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import { buildFilter, getFilterValue } from './use-view';
+import type { View } from '@wordpress/dataviews';
+import type { Moment } from 'moment';
+
+const EMPTY_ARRAY: ( ServerLog | PHPLog )[] = [];
+
+const useData = ( {
+	view,
+	logType,
+	dateRange,
+}: {
+	view: View;
+	logType: LogType;
+	dateRange: { startTime: Moment; endTime: Moment };
+} ) => {
+	const siteId = useSelector( getSelectedSiteId );
+	const severity = getFilterValue( view, 'severity' );
+	const status = getFilterValue( view, 'status' );
+	const requestType = getFilterValue( view, 'request_type' );
+
+	const { data, isFetching } = useSiteLogsQuery( siteId, {
+		logType,
+		start: dateRange.startTime.unix(),
+		end: dateRange.endTime.unix(),
+		filter: buildFilter( logType, severity, requestType, status ),
+		sortOrder: view.sort?.direction,
+		pageSize: view.perPage,
+		pageIndex: view.page,
+	} );
+
+	return {
+		data: data?.logs ? ( data.logs as ( PHPLog | ServerLog )[] ) : EMPTY_ARRAY,
+		paginationInfo: {
+			totalItems: data?.total_results || 0,
+			totalPages:
+				!! data?.total_results && !! view.perPage
+					? Math.ceil( data.total_results / view.perPage )
+					: 0,
+		},
+		isLoading: isFetching,
+	};
+};
+
+export default useData;
